requirements.txt
python-telegram-bot==21.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1

